Improve correctness of is_py2_stdlib_module() function
* This was not working before, which was causing problems with importing
* two modules such as:

    >>> from future import standard_library
    >>> standard_library.install_hooks()
    >>> import urllib.response
    >>> import urllib.request

After importing both of these (which scrubbed sys.modules in the process),
urllib.response no longer existed as an attribute of the ``urllib`` module
object.

A more robust solution might be to restore the state of the sys.modules cache at the
end of the hooks() and suspend_hooks() contexts, assuming there are no clashes
with names imported in the context.  If there are clashes, it would be good to
warn the user about this.