Remove BaseNewObject as a base class for other metaclass bases of future types.
It may be safer to remove BaseNewObject until we're sure it's needed because it
adds an extra layer of complexity.

However, this doesn't fix the tests introduced in f905bd for multiple
inheritance with other future types than ``newobject``; these still give a
``metaclass error``. So we mark the new tests as still failing.