Undocument support for EOL Python 2.6 and 3.3
Python 2.6 and 3.3 are end of life. They are no longer receiving bug
fixes, including for security issues. Python 2.6 went EOL on 2013-10-29
and on 2017-09-29. For additional details on support Python versions,
see:

https://devguide.python.org/#status-of-python-branches

Removing support for EOL Pythons will reduce testing and maintenance
resources.

Using pypinfo, here are the PyPI download statistics for the last 30
days, showing low numbers for these EOL Pythons:

| python_version | percent | download_count |
| -------------- | ------: | -------------: |
| 2.7            |  56.49% |      3,841,256 |
| 3.6            |  30.41% |      2,067,785 |
| 3.5            |   8.38% |        569,774 |
| 3.4            |   3.15% |        214,075 |
| 3.7            |   1.50% |        102,095 |
| 2.6            |   0.04% |          2,960 |
| 3.3            |   0.02% |          1,236 |
| 3.8            |   0.00% |             84 |
| 3.2            |   0.00% |             46 |
| None           |   0.00% |             33 |

https://github.com/ofek/pypinfo

This change only removes the documentation of 2.6 and 3.3. Legacy code
is expected to be removed in future commits as the changes are quite
extensive.