Bugfix with undefined children_hooks when package is None
touch_import_top() is always called with a package within
the futurize codebase, but it does have a code path where
package can be None (e.g. import six instead of from six
import string_types). Fixed a minor bug in this code path,
so that a custom fixer can use it.

Change-Id: Iec2891586fe852e35a91c69d2fb146645d7c53dd